+++ metadata.rb
@@ -3,11 +3,12 @@ maintainer       "Riot Games"
 maintainer_email "kallan@riotgames.com"
 license          "Apache 2.0"
 description      "Provides your cookbooks with the Artifact Deploy LWRP"
-version          "1.6.0"
+version          "1.9.0"
 
 supports "centos"
 supports "redhat"
 supports "fedora"
+supports "ubuntu"
 supports "windows", ">= 6.0"
 
 depends "windows", "~> 1.8.0"